Modifications
- Exit is now visible and controlled by SHOW_EXIT definition
- When the maze is solved it is shown in full in a green color
- Added a "We need to go back" phrase when encountering a Dead End
- Modified backtracking so it gets automatically to the last intersection
- Human can go on whatever direction they want independently of the status of the block

+++ Robobo.py
@@ -172,7 +172,7 @@ class Robobo:
         :param duration: Duration of the note in seconds (decimals can be used to used, like 0.2 or 0.5)
         :param wait: True: blocking mode, False: non-blocking mode
         :type note: int
-        :type duration: int
+        :type duration: float
         :type wait: bool
         """
         self.rem.playNote(note, duration, wait)
